refactor: split parties functionalities into dedicated file
